Cliffords simulation strategy added.
